H1/C1: extract PolyCore — the Qt-free operational core
New polyhost/core package (headless-core plan §2.1):
- poly_core.py: PolyCore owns the device stack (PolyKybd, DeviceManager,
  HidWorker + periodics), the reconnect probe, overlay send/cmd jobs,
  overlay mapping + handler, Sunlight, MRU persistence and the sleep
  listener. Results flow exclusively through observer callbacks
  (emit(name, payload)) documented in core/events.py — names deliberately
  match the GUI's existing job_done dispatch, so the Qt adapter is a
  single queued-signal hop. pywinctl is imported lazily and window
  tracking degrades to off without a display (plan §5.4).
- decisions.py: decide_probe_publish/decide_reconnect_apply moved
  verbatim out of gui/worker_bridge.py (which re-exports them).
- tests/core/import_guard_test.py: imports the core package in a fresh
  interpreter with PyQt5 poisoned — the headless contract from H1 on.

host.py shrinks accordingly: construction delegates to PolyCore,
connection state (connected/device_present/paused/last-applied) becomes
properties over the core's single source of truth, pause/quit/MRU/
overlay-send/window-tick delegate, and the moved probe/periodic/job
bodies are deleted. The reconnect APPLY (decision tree + menu/icon
updates) intentionally stays GUI-side for this commit — it moves in C2.
